Clean up correctly on exit
We need to ensure greenlets get killed in order to fire finally and exit blocks.
The lack of this was causing a problem where the term attrs were remaining set to raw mode.

As a nice other note, we now actually send a QUIT message instead of just hard closing the connection.